Moved scan / masslist sorter to util
