+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/spectraldbsubmit/view/MSMSLibrarySubmissionWindow.java
@@ -62,7 +62,6 @@ import net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.GnpsValues.Comp
 import net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.LibrarySubmitIonParameters;
 import net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.LibrarySubmitParameters;
 import net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.LibrarySubmitTask;
-import net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.sorting.ScanSortMode;
 import net.sf.mzmine.modules.visualization.spectra.multimsms.pseudospectra.PseudoSpectrumDataSet;
 import net.sf.mzmine.parameters.Parameter;
 import net.sf.mzmine.parameters.UserParameter;
@@ -74,6 +73,7 @@ import net.sf.mzmine.util.PeakListRowSorter;
 import net.sf.mzmine.util.SortingDirection;
 import net.sf.mzmine.util.SortingProperty;
 import net.sf.mzmine.util.components.GridBagPanel;
+import net.sf.mzmine.util.scans.sorting.ScanSortMode;
 
 /**
  * Holds more charts for data reviewing
