+++ src/main/java/net/sf/mzmine/util/ScanUtils.java
@@ -41,9 +41,9 @@ import net.sf.mzmine.datamodel.RawDataFile;
 import net.sf.mzmine.datamodel.Scan;
 import net.sf.mzmine.datamodel.impl.SimpleDataPoint;
 import net.sf.mzmine.main.MZmineCore;
-import net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.sorting.ScanSortMode;
-import net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.sorting.ScanSorter;
 import net.sf.mzmine.util.exceptions.MissingMassListException;
+import net.sf.mzmine.util.scans.sorting.ScanSortMode;
+import net.sf.mzmine.util.scans.sorting.ScanSorter;
 
 /**
  * Scan related utilities
